Upgrade Node and NPM versions
We are upgrading our Cloud Foundry from v245 to v250, which means
the nodejs-buildpack will now be version 1.5.24, which does not support
Node v6.7.0. We configure it with a semantic version, which means the buildpack
will pick the highest version it supports. We do the same for the npm version,
otherwise the pinned npm version will eventually become out of date.