Add card toogle for each team
We've split into two teams. This means, we've got two separate standups.
In order to make it easy for our selves and not go through every card
thinking it may be a team specific, we'd like to have an option, to hide
the cards not relevant to us.

This change introduces an option bar, which has ability to filter the
cards by the team.