Add token header for each state
Instead of, rewriting the rubbernecker to use websockets or ajax
functionality, we could implement a little token checkup.

Here, we combine the most important information about the current board
state and generate a SHA512 as a token, to be returned to the client.

It also sketched out, the possibility to check on the token, by
comparing the one given by the client, with current on hand. It should
return a json content, claiming that the token is `clean` or `dirty`.