Implement state auto changes on the front end
In order, not to be out of date, we'd like to implement a little check,
that would be running in the background of the application and compare
the tokens provided of the content obtained from Pivotal.

If the change is detected, we should get a little notice, mentioning
that, with a countdown giving us chance to cancel or manually refresh
the page.

Option, to control auto-update has been added into option bar.