Add support for the semi-query string
As per feature request, we'd like to be able to maintain our choices
inside of the url.

In order, to keep the JS management, this URL needs to be based on HASH.
This should allow us, to change the URL as we change the options and
then keep these options, on refresh.

A lot of syntax, had to be modified to get everything to work as
expected...