Attempt to recover
There is no reason to kill the application if something fails to load.
We can attempt to reload the page every ten seconds in order to
establish a new connection.

This should only cover the errors we're able to parse. If the CF fails
over or the server is not responding with any body, it will still
freeze.