Fix the number of days in play
The current approach is, that we do not care for stories that took
longer than two weeks. WRONG! We've got more of these recently, and it
would be useful, to have an actual corresponging number.

It's a little bit of a pain, to calculate these correctly. I took the
approach, of doing that after all the stories have been collected. Each
of the story, then will call the API for it's own transitions over time
and the days will be calculated from that. It means, there is a slight
increase of API, CPU and Memory usage. It also increases the
rubbernecker response time. I am not proud about any of these, but it's
a terrible hack we have to live with, if we'd like to be precise.

This commit, also covers some minor clean up and change of some
approaches.