Check if the transition exists
The Pivotal API returns an array of transitions, which is then used to
calculate number of days in current state. It would appear, it also
returns a `null` value, which simply kills our application.

We're now checking if the value returned by an API is not `null` nor
undefined type.